Fix P2 UltimateDeployment: health cache race, blue-green rollback, canary steps, int division, K8s TOCTOU
- DeploymentStrategyBase: _cachedHealthy/_healthCacheExpiry non-volatile → _healthCacheLock (finding 2868)
- BlueGreenStrategy: rollback reads 'environment' key that Deploy never writes; derive
  logical env name by stripping -blue/-green suffix from activeEnv (finding 2873)
- CanaryStrategy: traffic steps were not monotonically increasing; fixed with
  Where(s > canaryPercent).Prepend(canaryPercent) (finding 2875)
- InfrastructureStrategies: integer division truncated progress to 0 for all but last
  host; fixed to float division (finding 2887)
- KubernetesCsiDriver: capacity reservation CAS loop (finding 2899), SingleNodeWriter
  TOCTOU moved check inside lock (finding 2900), ListVolumes NextToken fixed to
  (skip+maxEntries).ToString() (finding 2898)
- UltimateDeploymentPlugin: DeployAsync stamps strategy name into metadata so
  RollbackAsync can find the correct strategy (finding 2870)

+++ Plugins/DataWarehouse.Plugins.UltimateDeployment/Strategies/VMBareMetal/InfrastructureStrategies.cs
@@ -704,7 +704,7 @@ public sealed class SshDirectStrategy : DeploymentStrategyBase
 
         foreach (var host in hosts)
         {
-            state = state with { ProgressPercent = 20 + (60 * Array.IndexOf(hosts, host) / hosts.Length) };
+            state = state with { ProgressPercent = (int)(20 + (60.0 * Array.IndexOf(hosts, host) / hosts.Length)) };
 
             // Connect via SSH
             await SshConnectAsync(host, ct);
